# Optimize VectorWriter<Varchar> for up to 2X faster for some string functions