--- a/velox/expression/StringWriter.h
+++ b/velox/expression/StringWriter.h
@@ -29,20 +29,9 @@ class StringWriter;
 template <>
 class StringWriter<false /*reuseInput*/> : public UDFOutputString {
  public:
-  StringWriter() : vector_(nullptr), offset_(-1) {}
-
   // Used to initialize top-level strings and allow zero-copy writes.
   StringWriter(FlatVector<StringView>* vector, int32_t offset)
       : vector_(vector), offset_(offset) {}
-
-  // Used to initialize nested strings and requires a copy on write.
-  /* implicit */ StringWriter(StringView value)
-      : vector_(nullptr), offset_(-1), value_{value.str()} {}
-
-  // Returns true if initialized for zero-copy write. False, otherwise.
-  bool initialized() const {
-    return offset_ >= 0;
-  }
 
   // If not initialized for zero-copy write, returns a string to copy into the
   // target vector on commit.
@@ -126,7 +115,6 @@ class StringWriter<false /*reuseInput*/> : public UDFOutputString {
 
   template <typename T>
   void copy_from(const T& input) {
-    VELOX_DCHECK(initialized());
     append(input);
   }
 
@@ -135,6 +123,8 @@ class StringWriter<false /*reuseInput*/> : public UDFOutputString {
   }
 
  private:
+  StringWriter() = default;
+
   bool finalized_{false};
 
   /// The buffer that the output string uses for its allocation set during
@@ -146,6 +136,9 @@ class StringWriter<false /*reuseInput*/> : public UDFOutputString {
   int32_t offset_;
 
   std::string value_;
+
+  template <typename A, typename B>
+  friend struct VectorWriter;
 };
 
 // A string writer with UDFOutputString semantics that utilizes a pre-allocated