--- a/velox/functions/prestosql/benchmarks/CMakeLists.txt
+++ b/velox/functions/prestosql/benchmarks/CMakeLists.txt
@@ -124,3 +124,8 @@ add_executable(velox_functions_benchmarks_row_writer_no_nulls
                RowWriterBenchmark.cpp)
 target_link_libraries(velox_functions_benchmarks_row_writer_no_nulls
                       ${BENCHMARK_DEPENDENCIES_NO_FUNC})
+
+add_executable(velox_functions_benchmarks_string_writer_no_nulls
+               StringWriterBenchmark.cpp)
+target_link_libraries(velox_functions_benchmarks_string_writer_no_nulls
+                      ${BENCHMARK_DEPENDENCIES_NO_FUNC})