Summary:
- This mostly affects simple functions that write strings:
- substring and array_of_strings are 2X faster after this change.

- A follow-up diff will change vector functions that use the
StringWriter to use  VectorWriter<Varchar> in the vector function to
gain the same benefits.
```
From StringWriterBenchmark.cpp
not_inline_string
 58.50ns  -> 28.53ns

inline_string
 46.00ns ->  28.53ns

array_of_string 12.58ms
618.39ns  -> 238.68ns

From StringAsciiUTFFunctionBenchmarks.cpp
(opts.stringLength = 100, opts.vectorSize = 1000):

asciiSubStr
23.63ms -> 12.61ms (2X)

asciiRPad
  5.01ms -> 3.06ms

aciiLPad
 4.66ms -> 2.94 ms
```

- The main optimization is avoiding reconstructing the proxy and
recomputing the buffer, capacity and locations to write to for the
every row, but instead utilizes previous state.

Differential Revision: D36356103